Echo crypto suite tag of the offer SDP in the answer SDP
Currently, when Livekit generates an answer SDP, the crypto suite will
have a tag value of whatever index was assigned to it by
``DefaultProfiles()``. This is a problem in case the offerer strictly
validates that the tag in the answer is equal to the one in the offer.

RFC 4568 section 5.1.3:

```
When the offerer receives the answer, the offerer MUST verify that
one of the initially offered crypto suites and its accompanying tag
were accepted and echoed in the answer
```

This patch should solve the last of the SDES-related compatibility issues
with rtpengine that I encountered.

Note that this does not address the fact that Livekit ignores the crypto
suite tag when processing an answer SDP.